Link revenue and appointment to ui
