+++ src/main/resources/view/appointmentpanel/AppointmentListPanel.fxml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.ListView?>
+<?import javafx.scene.layout.VBox?>
+
+<VBox xmlns="http://javafx.com/javafx/11" xmlns:fx="http://javafx.com/fxml/1">
+    <ListView fx:id="appointmentListView" VBox.vgrow="ALWAYS" />
+</VBox>
