+++ src/main/resources/view/revenuepanel/RevenueListPanel.fxml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.ListView?>
+<?import javafx.scene.layout.VBox?>
+
+<VBox xmlns="http://javafx.com/javafx/11" xmlns:fx="http://javafx.com/fxml/1">
+    <ListView fx:id="revenueListView" VBox.vgrow="ALWAYS" />
+</VBox>
