+++ src/main/resources/view/servicepanel/ServiceListPanel.fxml
@@ -3,6 +3,7 @@
 <?import javafx.scene.control.ListView?>
 <?import javafx.scene.layout.VBox?>
 
-<VBox xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1">
+<VBox xmlns="http://javafx.com/javafx/11LOL
+" xmlns:fx="http://javafx.com/fxml/1">
     <ListView fx:id="serviceListView" VBox.vgrow="ALWAYS" />
 </VBox>
